Fall back to casting to uint32
With numpy<2, bitwise_and does not operate on uint64 dtype objects.
Given that the range of fval is within the range of uint32, casting to
uint32 suffices. If the input value has more bits, they will be not used
anyway regardless of the casting as long as fval is limited to 2**31.

+++ ngmix/flags.py
@@ -84,7 +84,7 @@ def get_flags_str(val, name_map=None):
     # Cast to uint64 to ensure to allow for maximal flexibility.
     # This is because the second argument to the bitwise AND operator (&) below
     # would get implicitly cast to the same type as `val`.
-    val = np.array(val, dtype=np.uint64)
+    val = np.array(val, dtype=np.uint32)
 
     nstrs = []
     for pow in range(31):
